pyproject.toml
[project]
name = "hajimi-king"
version = "0.0.1-beta"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "pylint>=3.3.8",
    "ruff>=0.12.9",
    "pre-commit>=4.0.0",
]

[tool.ruff]
# Exclude patterns
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "*.egg-info",
    "build",
    "dist",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
]

line-length = 120
target-version = "py311"

[tool.ruff.lint]
# Enable specific rule categories
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "PIE", # flake8-pie
    "SIM", # flake8-simplify
    "RET", # flake8-return
    "ERA", # eradicate (commented code)
]

# Ignore specific rules if needed
ignore = [
    "E501",  # line too long (handled by formatter)
    "UP007", # Use X | Y for type annotations (Python 3.10+ syntax)
    "ERA001", # eradicate (commented code) - too aggressive for Chinese comments
    "RET504", # unnecessary assignment before return
    "SIM117", # nested with statements
    "SIM105", # contextlib.suppress
    "SIM112", # capitalized env vars
    "SIM102", # nested if statements
    "SIM108", # ternary operators
    "SIM103", # return condition directly
    "B007",  # unused loop variables
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py files

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Use spaces around = in keyword arguments
skip-magic-trailing-comma = false

[tool.ruff.lint.flake8-import-conventions.aliases]
# Standard library aliases
"datetime" = "dt"
"matplotlib.pyplot" = "plt"
"numpy" = "np"
"pandas" = "pd"
